+++ backend/models/Bill.js
@@ -12,7 +12,7 @@ module.exports = (sequelize,Sequelize) =>{
           
           },
           amount: {
-            type: Sequelize.DECIMAL(10, 2), // Assuming decimal for amount, adjust as necessary
+            type: Sequelize.DOUBLE(10, 2), // Assuming DOUBLE for amount, adjust as necessary
             allowNull: false
           },
